슬라이드 셰어 임베드 가능하게 변경
